LLMOPS/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixit-llmops"
version = "0.1.0"
description = "Local LLMOps reference implementation for the FixIt AI Support System."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.135.0",
  "openai>=2.31.0",
  "pydantic>=2.13.0",
  "PyYAML>=6.0.0",
  "uvicorn>=0.44.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

